refactor: game-logic.ts のドメイン分割（P2）
- game-logic.ts（1,356行）を16のドメインサービスファイルに分割
- domain/ 配下に shared/battle/evolution/skill/awakening/event/progression/achievement/challenge の9ドメインを構築
- game-logic.ts を barrel re-export ファイルに変換（後方互換性維持）
- 16のドメインテストファイルを TDD で作成（計1,232行）
- RIT_LOW_HP_RATIO の重複定義を combat-calculator.ts に一元化（DRY原則）
- deepCloneRun/writeSnapToRun をドメイン内共有ユーティリティとして配置
- 全テストパス（245スイート / 3,320テスト）、型エラーなし、ビルド成功
